feat(api-server): add ability to get k8s format of a resource
For any GET call an optional parameter `?format=k8s` can be
passed to get the resource as a k8s resource.
You can also use `?namespace=<whatever>` to use a override the
namespace.

This aims to have the UI always show universal resources and only
provide a `copy as k8s` button that does this call.

Fix #6658

Signed-off-by: Charly Molter <charly.molter@konghq.com>