fix(kuma-cp): make store changes processing more reliable (#6728)
When running deployment with postgres/etcd we are using EventBus which is responsible for sending database events to mesh-insight-resyncer component. In case of a problem with the connection to the database, mesh-insight-resyncer component is closed and restarted by ResilientComponent. On each restart, we are subscribing to the event bus, but we are not removing the old subscription. That caused the issue in which we were sending events to the subscriber that didn't exist at that time. Because we are sending it and no one is listing it, the component was hanging and events were not propagated to other subscribers.

Signed-off-by: Lukasz Dziedziak <lukidzi@gmail.com>