chore(policy/MeshTCPRoute): make to[].rules optional
There is currently no validation if the
`to.rules[].default.backendRefs` doesn't direct the traffic to
the same target as matched in `to.targetRef`.
It means the configuration as below is valid:
```yaml
targetRef:
  kind: MeshService
  name: frontend
to:
- targetRef:
    kind: MeshService
    name: backend
  rules:
  - default:
      backendRefs:
      - targetRef:
          kind: MeshService
          name: backend
```
Why are we not disallowing it during the validation you might ask?
We don't as:
- it's non-breaking to have a route which direct the traffic
  to the same destination
- it would involve in doing very heavy validation otherwise (we
  would have to compute potential splits and check it summarized
  backendRefs don't point to the same destination etc.)
- there is assumption that route-policies like `MeshTCPRoute` will
  in the near future allow to target `MeshGateway's as a root
  targes, which makes this structure as perfectly valid

Because of reasons stated above, I belive there is no need to
enforce specification of the rules, and this PR makes this field
optional, so the example above might be simplified to:
```yaml
targetRef:
  kind: MeshService
  name: frontend
to:
- targetRef:
    kind: MeshService
    name: backend
```

As now you'll be able to apply policy without specifying `to.rules`
and as there is no matching capabilities for tcp routes every
rule should contain at least one `backendRef` so `backendRefs`
field is now required to contain at least 1 `backendRef`.

Signed-off-by: Bart Smykla <bartek@smykla.com>