refactor(MeshHTTPRoute): pull up parts of xds generation
As for `MeshTCPRoute` endpoints and clusters generation will be the same
instead of copying this code, I took it out and made it reusable.

I had also changed `GetClusterName` definition, to be reusable, so
instead of taking concrete targetRefs from `MeshHTTPRoute`, now we need
to provide `name` and `tags` as separate arguments.

Signed-off-by: Bart Smykla <bartek@smykla.com>