refactor(MeshHTTPRoute): extract BackendRefs
`BackendRef` is a structure which in the same form can be used in other
route policies, like `MeshTCPRoute`. Reusing the same structure also
allows us to extract function `MakeHTTPSplit`, which uses it to build
splits used in building Envoy listeners.

As the file now contains not only `TargetRef`, but also `BackendRef`,
I changed the name of it from `backendref.go` to `ref.go` which
I believe is now more appropriate (even if `BackendRef` structure
is very small).

Signed-off-by: Bart Smykla <bartek@smykla.com>