refactor(MeshHTTPRoute): cluster in svc accumulator with summed weights
When adding clusters to service accumulator now we are assigning a
weight which is a sum of all weights from refs pointing to the same
destination.

Having weight in cluster doesn't affect splits which are used by
`MeshHTTPRoute`. Services in service accumulator will also contain
the same amount of clusters. The only difference will be the clusters
will now contain appropriate weights, which might be used in other route
policies like `MeshTCPRoute`.

Signed-off-by: Bart Smykla <bartek@smykla.com>