fix(kuma-cp): make zone insight context independent from parent (#6909)
When we are using parent context and the stream is canceled function responsible for flushing stats to the database is also canceled and fails. We need a separate context that still can flush data after the parent context is closed but the stream is not fully destroyed. That happens when using pgx and postgres.

Signed-off-by: Lukasz Dziedziak <lukidzi@gmail.com>