feat(api-server): make errors compliant with aip 193
We've decided to start following for our apis: https://kong-aip.netlify.app/aip/193/

This PR makes errors return both the old and new fields for back
compatibility.
In a future version we'll be able to remove old fields

We also have error responses in golden files for easier changes.

Fix #6136

Signed-off-by: Charly Molter <charly.molter@konghq.com>