ci(docs): add `make docs/output` to copy all raw docs (#7146)
The goal of this is to take the configuration files (crds, proto json
schema, yaml confs) with no modification and get them in a directory.

These will be then used in the docs website where they can be modified
as needed. That way the logic stays in one place (the website).

Signed-off-by: Charly Molter <charly.molter@konghq.com>