fix(kuma-dp): pass sockets in metadata from dp to cp
Currently socket paths where generated from both the cp and dp.
These relied on `os.TempDir()` which may return different results as
they run on 2 different machines.

With this change we make the paths configurable from the kuma-dp
process.
The information is then sent in the metadata and can be used during
config generation.

From a backward compatibility POV there are 2 things to take care of:

- We may update the CP and have an old DP still running (the metadata
  won't have the entries for sockets).
- We may have an old DP bootstrap on a newer CP (this bootstrap request
  won't have the DP set).

Fix #5228

Signed-off-by: Charly Molter <charly.molter@konghq.com>