chore(GlobalInsight): allow using GlobalInsightService in other places
Allow using GlobalInsightService in other than API endpoint places

- Modified GlobalInsightService to use ResourceStore instead of
  ResourceManager, as it didn't use any specific ResourceManager
  functionality.
- Exposed GlobalInsightService in the Runtime and RuntimeBuilder
  interfaces.
- Fixed the returned type of GlobalInsight endpoint to be JSON
  instead of text.

Explanation:

The GlobalInsightService was previously only accessible from
API endpoint places. To allow it to be used in other places,
such as Bootstrap plugins, it needed to be modified to use
ResourceStore instead of ResourceManager.

In addition, the GlobalInsightService needed to be exposed
in the Runtime and RuntimeBuilder interfaces. This makes it
available to a wider range of code, including Bootstrap plugins.

Finally, the commit fixes the returned type of the GlobalInsight
endpoint to be JSON instead of text. This is because JSON is
the standard format for data exchange between web applications.

Overall, this commit makes the GlobalInsightService more accessible
and easier to use.

Signed-off-by: Bart Smykla <bartek@smykla.com>