refactor(xds): split MeshContext in 3 to be able to improve caching
Currently MeshContext contains everything. Which means that
whenever something changes on the mesh we pull a lot of things from
the DB.

In practice there are 3 types of resources:

- global resources zone proxies, meshes...
- zone resources that rarely change policies, externalServices ...
- the rest which changes often dataplanes

We now have 3 different things built independently. This will enable us
to:

- Not pull down policies when a dataplane changes if we make things
  event based.
- Change cache invalidation speed depending on the type (Global stuff
  could be cached longer or we could decide to propagate policy changes
less quickly to improve caching).
- Write inspect APIs without getting the full mesh context (only the
  policies are useful).

In its current implementation we're not expecting any strong perf
improvements as we don't use the new context independently

Also use fnv128 instead of sha256. The probability of collision
is low and using a non crypto hash will be quicker

Signed-off-by: Charly Molter <charly.molter@konghq.com>