fix(k8s): fix VIPs configmap entries with invalid keys for ExternalName services (#8168)
Services with the `ExternalName` type do not allow us to get an
IP address, so the generated VIPs config map contains entries with
invalid keys, such as `1:`. This was hard to debug because we were
getting an EOF error during unmarshalling of the VIPs config map.

This commit fixes the issue by ignoring services of that type with
appropriate log when it'll happen.

Signed-off-by: Bart Smykla <bartek@smykla.com>