feat(kuma-dp): add coredns logging flag
As somebody who is trying to set up a mesh in kubernetes and run terraform in it,
I need a way to turn on CoreDNS query logging or debugging on, so that I can use
the logs from that to enumerate all the external services that I need to allow
outbound access to.

This patch adds a flag to the helm chart that will inject an environment variable
into the dataplane sidecars that will add the "log" plugin to the CoreDNS config.
I added a test to make sure the config would get set in dnsserver, and I checked
that `helm template` would add the `KUMA_RUNTIME_KUBERNETES_INJECTOR_BUILTIN_DNS_LOGGING`
env var to the control plane, which will cause the kuma-dp sidecars to get
logging turned on and going to stdout.

I also ran `make check` and `make test` cleanly.

Fix https://github.com/kumahq/kuma/issues/8254 and maybe https://github.com/kumahq/kuma/issues/5516

Signed-off-by: Tim Spencer <timothy.spencer@gsa.gov>