chore(zoneproxies): don't fail when mesh was deleted
To build the aggregateMeshContext we first list the meshes
and then retrieve the MeshContext. It's possible that between
these 2 operations a mesh was deleted.
We simply ignore these now, the worst that can happen is that
a mesh is ignored for a short time from the Egress which seems
unlikely.

This should suppress a lot of error logs on the CP

Signed-off-by: Charly Molter <charly.molter@konghq.com>