chore(MeshProxyPatch): handle multiple modification types in a single object
Previously, the MeshProxyPatch policy only validated the first
element in the provided modification object. This meant that if
a modification object contained multiple types of modifications
(e.g., cluster, listener and networkFilter), only the first type of
modification would be applied, and the others would be silently
ignored. To address this, I updated the validator to allow multiple
modification types to be specified within a single modification
object. This enables users to define more complex modifications
in a single place, without having to split them across multiple
objects.

Signed-off-by: Bart Smykla <bartek@smykla.com>