fix(kds): avoid rare cases where onStreamClosed is called with no state
There were rare cases where state is not in the streams map.
We simply avoid a nil pointer by returning early in this case.
There shouldn't be any problems with this as the entry is not in the map
anyway

However, maybe there's a race where streamClosed is called before
streamOpen

Signed-off-by: Charly Molter <charly.molter@konghq.com>