fix(kumactl): kumactl -o yaml breaks Mesh* policies
After the release of the new Mesh* policies, fixing issue #679 broke
the output of content in YAML format. The fix for issue #679, which
involves splitting the policy into a header and spec body and
outputting them separately, resulted in the loss of the spec in this
particular case. This issue can be resolved by using an alternative
method for converting YAML to JSON.

Signed-off-by: Nikolay Neborsky <nick.nnn@gmail.com>