test(e2e/framework): don't fail when deleting non-existing k8s namespace
When removing kic deployment we are removing a configured ingress
namespace. There might be situation when this namespace might be
already deleted during different test suite stage, so if doesn't
exist, we don't have to fail during the deletion.

Updated kic deployment delete function to call `TraiggerDeleteNamespace`
(which potentially might be more performant) instead of
`DeleteNamespace`.

Signed-off-by: Bart Smykla <bartek@smykla.com>