fix(gatewayapi): validate presence of all necessary gapi resources
Existing check of presence for only `Gateway` resource is
insufficient in situation when some (not all) of the Gateway API
CRDs are missing (i.e. package with CRDs was installed before
`ReferenceGrant` was added), which causes CP to start throwing
lot of errors:

```
controller-runtime.source.EventHandler	if kind is a CRD, it should
be installed before calling Start	{"kind":
"ReferenceGrant.gateway.networking.k8s.io", "error": "no matches
for kind \"ReferenceGrant\" in version
\"gateway.networking.k8s.io/v1beta1\""}
```

This is the simplest solution as we could also verify if
`ReferenceGrant` is present, and if not to disable cross-mesh
references, but it would be unnecesarily complex, especially
in situation when Gateway API is GA and contains `ReferenceGrant`
in `v1`.

Signed-off-by: Bart Smykla <bartek@smykla.com>