chore(api-server): pass runtime extensions to endpoint handlers
When using `AddFieldFromCtx` in `HandleError` function we were
passing fresh, empty context. Because of that whenever logger
had custom `FromSpanLogValuesProcessorContext` passed, it was
ignored. Now we are passing runtime extension context to all
endpoint handlers which then are using `HandleFunction`, which
allows us to provide in example mapper of OTEL trace ids to
Datadog.

Signed-off-by: Bart Smykla <bartek@smykla.com>