Rename "commands" to "rules"
Rules are more flexible if we want to introduce some additional
logic for example to remove particular iptables rule or check
if it exists. Before we would end up with a list of commands
inside chains, which would make it harder if we in example would
like to check if the rules we want to add/remove already exists,
because we would have to iterate over all "commands" inside chains
and create new `Command`s for check (-C|--check) or delete.

Signed-off-by: Bart Smykla <bartek@smykla.com>