Enhance InitializedExecutable to support default arguments
The `args` field in `InitializedExecutable` now stores a set of default
arguments or flags that are consistently included in every execution of
the executable. These default arguments are prepended to any additional
arguments passed in the Exec method. This enhancement ensures that
specific flags or options are consistently applied whenever the
executable is invoked.

Signed-off-by: Bart Smykla <bartek@smykla.com>