feat(transparent-proxy): add iptables logging with new flag and annotation (#10743)
This commit introduces the ability to enable iptables logging for
packets matching the rules in our transparent proxy. This feature
enhances observability by logging details of the IP/IPv6 headers of
these packets, which can be crucial for debugging complex network
issues.

- New flag: `--iptables-logs` enables iptables logging during
  installation.
- New annotation: `traffic.kuma.io/iptables-logs` allows configuration
  via deployments, including Kubernetes init containers and CNI.

This logging is performed using the LOG chain and provides valuable
insights accessible through `dmesg(1)` or syslog. It records information
about packets that match iptables rules, helping in the analysis and
troubleshooting of network traffic.

Note: While enabling this feature can significantly aid in debugging, it
may also introduce a performance overhead due to the additional logging
activity. It is recommended to enable this feature judiciously,
considering the potential impact on system performance.

Signed-off-by: Bart Smykla <bartek@smykla.com>