fix(transparent-proxy): fix IPv6 iptables rules when no IPv6 DNS servers
When the `--redirect-dns` flag is present during the execution of
`kumactl install transparent-proxy`, the `/etc/resolv.conf` file is
parsed to obtain DNS servers. However, when the transparent proxy is
configured to include IPv6 and no valid IPv6 servers are found,
incorrect iptables rules were generated.

This commit fixes this when in such situation logs warning and skip
IPv6 DNS traffic redirection.

Signed-off-by: Bart Smykla <bartek@smykla.com>