fix(transparent-proxy): don't add comments if no comment iptables module (#10811)
In the PR introducing comments in iptables rules, I neglected to verify
the availability of the comment iptables extension/module. This change
addresses that oversight.

I acknowledge that passing the ipv6 value everywhere is cumbersome. I
plan to refactor the builders to handle IP types in a more agnostic
manner. This improvement will be addressed in subsequent PRs, as it
requires a more complex update.

Signed-off-by: Bart Smykla <bartek@smykla.com>