fix(transparent-proxy): use found executable path
There was a small bug where we were using provided path in findPath
function instead of the result of it which is a found path. This bug
shouldn't impact anyone as both cases are basically correct, but
assumption was to use found path, so if we were looking for `iptables`
we should use a found path which is a result of this function instead of
passed `iptables` as a path.

Signed-off-by: Bart Smykla <bartek@smykla.com>