fix(transparent-proxy): check DNS related CLI flags earlier
If both values `redirect.dns.enabled` and `redirect.dns.captureAll`
would be provided in the config yaml via `--config` or `--config-file`,
so far we would fail, but it's a correct scenario. We should check if
`--redirect-dns` and `--redirect-all-dns-traffic` are not provided
together, so we should do it before we'll load config from file.

Signed-off-by: Bart Smykla <bartek@smykla.com>