feat(transparent-proxy): allow to use tproxy config from configmap in kubernetes (#11403)
This update enables Kubernetes deployments to use the transparent
proxy feature, allowing the configuration to be provided via
the `--config` or `--config-file` CLI flags. When
`--set transparentProxy.configMap.enabled=true` is set,
the configuration will be stored in a ConfigMap. Additionally, you can
use the `traffic.kuma.io/transparent-proxy-configmap-name`
annotation to specify a different ConfigMap for specific workloads.

Signed-off-by: Bart Smykla <bartek@smykla.com>