chore(mads/hds): simplify snapshot versioning
We were having flakes in the MADS test due to snapshots
getting a different version even though they were the same.

The snapshot versioning code was unecessary complicated.
This was rewritten in a simpler manner which removes the flakes.

Improved the mads tests to make there more readable and easier to track
issues

Signed-off-by: Charly Molter <charly.molter@konghq.com>