ci(e2e): run `K8S_MIN_VERSION` tests on arm64
We have a lot of arm runners and starve amd runners.
We have very high trust in our arm support so comparing
amd and arm on the same test suite is a little wasteful.

We therefore only run e2e tests on old versions of k8s
on arm64 this removes a few entries in the matrix and
improve utilization of arm runners

Signed-off-by: Charly Molter <charly.molter@konghq.com>