chore(mads/hds): remove custom cache implementation
In mads and hds we were using a custom implementation of a cache.
This was unfortunate as go-control-plane has a perfectly working one.

The only feature on this cache was the ability to long poll.
This was only required in mads and we've changed the handler of mads
to make this handling outside of the cache.

Signed-off-by: Charly Molter <charly.molter@konghq.com>