fix(kuma-cp): remove automatically created MeshServices when mode is switched to `Disabled` (#11675)
Users should have a safe way to rollback the MeshService migration. Today switching from Disabled to Exclusive to Disabled leads to a situation when MeshServices are not deleted.

There are 2 changes for Kubernetes and Universal. When mode is Disabled we remove automatically created MeshServices. The alternative would be to keep generated MeshServices when mode is switched to Disabled, but check the mode on mesh every time we generate Listeners and Clusters. That creates another potential state of the cluster that we should take into account. The cleanest approach is to rollback to the exact state user had before the migration.

Signed-off-by: Ilya Lobkov <ilya.lobkov@konghq.com>