chore(transparent-proxy): deprecated code cleanup and small improvements (#11732)
* Remove deprecated CLI flag and config --redirect-dns-upstream-target-chain

It was not used and in deprecation message it was mentioned it will be
removed in kuma 2.9

* Ensure all fields in `DefaultConfig` are explicitly set for consistency

This update initializes all fields in `DefaultConfig`, even those with
zero values, for consistency. Now, you can easily inspect the full
structure of `Config` by reviewing the `DefaultConfig` function without
needing to dig through nested structures. Previously, some fields had
zero values assigned, while others did not, leading to inconsistency.
This change standardizes the initialization approach across all fields.

* Refactor jump parameters + cleanup some constants

Rename `JumpParameter` to `WrappingParameter` and move it out to
`parameters.go` so it can be reused

Refactor `Log` to take `...*WrappingParameter` to configure it, so in
the future if we'll decide to add more flags fitting this jump target,
it'll be easier to do

Cleanup some constants and remove unnecessary ones

* Introduce `InitializedRetry` structure, which simplifies usage of `Retry`

Signed-off-by: Bart Smykla <bartek@smykla.com>