test(e2e): disable amd64 self-hosted runners (#11862)
IPV6 tests seems to be broken with our self-hosted runners.
It's not clear what's the problem, it can take some time to resolve it.
We'd rather have a slower CI with queues than CI which is broken. We can
still keep arm64 tests on self-hosted runners because those do not run
IPV6.

Signed-off-by: Jakub Dyszkiewicz <jakub.dyszkiewicz@gmail.com>