chore(mads/hds): remove custom cache implementation (#11590)
In mads and hds we were using a custom implementation of a cache. This
was unfortunate as go-control-plane has a perfectly working one.

The only feature on this cache was the ability to long poll. This was
only required in mads and we've changed the handler of mads to make this
handling outside of the cache.

This is more complete fix to problems like:
https://github.com/kumahq/kuma/pull/11578 as there is now only one
reconciling loop.

### Checklist prior to review

<!--
Each of these sections need to be filled by the author when opening the
PR.

If something doesn't apply please check the box and add a justification
after the `--`
-->

- [x] [Link to relevant issue][1] as well as docs and UI issues --
- [x] This will not break child repos: it doesn't hardcode values (.e.g
"kumahq" as a image registry) and it will work on Windows, system
specific functions like `syscall.Mkfifo` have equivalent implementation
on the other OS --
- [x] Tests (Unit test, E2E tests, manual test on universal and k8s) --
  - Don't forget `ci/` labels to run additional/fewer tests
- [x] Do you need to update [`UPGRADE.md`](../blob/master/UPGRADE.md)?
--
- [x] Does it need to be backported according to the [backporting
policy](../blob/master/CONTRIBUTING.md#backporting)?
([this](https://github.com/kumahq/kuma/actions/workflows/auto-backport.yaml)
GH action will add "backport" label based on these [file
globs](https://github.com/kumahq/kuma/blob/master/.github/workflows/auto-backport.yaml#L6),
if you want to prevent it from adding the "backport" label use
[no-backport-autolabel](https://github.com/kumahq/kuma/blob/master/.github/workflows/auto-backport.yaml#L8)
label) --

<!--
> Changelog: skip
-->
<!--
Uncomment the above section to explicitly set a [`> Changelog:` entry
here](https://github.com/kumahq/kuma/blob/master/CONTRIBUTING.md#submitting-a-patch)?
-->

[1]:
https://docs.github.com/en/issues/tracking-your-work-with-issues/linking-a-pull-request-to-an-issue#linking-a-pull-request-to-an-issue-using-a-keyword

---------

Signed-off-by: Charly Molter <charly.molter@konghq.com>