fix(cni): delegated gateway was not correctly injected (#11922)
I wanted to run Zone 2 in e2e tests in CNI mode and I encountered a
problem with injection.
We were injecting such init container
```
  initContainers:
  - name: ""
    resources: {}
```
I fixed it and switched one zone to CNI to avoid this in the future.

Inject container only if it's created

Placing run-full-matrix, because I need to see if all variants are ok. I
checked locally and e2e is fine with CNI.

Signed-off-by: Jakub Dyszkiewicz <jakub.dyszkiewicz@gmail.com>