feat(openapi): add creationTime and modificationTime to policy specs (#12006)
In order to further complete the openAPI specs, we need to add
`creationTime` and `modificationTime`. These properties are already part
of the response payload (see context below). Since the GUI project
started to generate the TypeScript types from the specs, there is a high
demand in having complete specs to avoid manual backfilling.

Signed-off-by: schogges <moritz.fleck@konghq.com>