fix(meshtimeout): don't set default timeouts on inbound cluster and listener (#12043)
We noticed that when a user upgrades to version 2.9.x without a default
MeshTimeout/Timeout policy, default timeouts are applied on inbound
clusters and listeners. This could disrupt the existing behavior of the
environment.

When no policy is present, we simply skip configuration and do not set
up the timeouts for inbound cluster or listener.

Fix https://github.com/kumahq/kuma/issues/12033

<!--
> Changelog: skip
-->
<!--
Uncomment the above section to explicitly set a [`> Changelog:` entry
here](https://github.com/kumahq/kuma/blob/master/CONTRIBUTING.md#submitting-a-patch)?
-->

---------

Signed-off-by: Lukasz Dziedziak <lukidzi@gmail.com>