test(e2e): change consitently to eventually (#12097)
## Motivation

I've noticed a flake in a specific test. We are using `Consistently`,
which expects that error code `52` will always be returned. This is a
correct assumption, unless something happens on the network and we
receive a different error code.

```
  [FAILED] Failed after 0.250s.
  The function passed to Consistently failed at github.com/kumahq/kuma/test/e2e/reachableservices/auto_reachable_services_k8s.go:100 with:
  Expected
      <int>: 56
  to equal
      <int>: 52
Error: It 11/22/24 10:02:24.833
```

## Implementation information

To avoid this flake, I changed `Consistently` to `Eventually` and added
`MustPassRepeatedly`. We cannot guarantee that the network is 100%
reliable, so we have two options:

* Add another error code to check, or
* If it fails, wait for 5 consecutive successes.

## Supporting documentation

https://github.com/kumahq/kuma/actions/runs/11969768862/job/33372432545

<!--
> Changelog: skip
-->
<!--
Uncomment the above section to explicitly set a [`> Changelog:` entry
here](https://github.com/kumahq/kuma/blob/master/CONTRIBUTING.md#submitting-a-patch)?
-->

Signed-off-by: Lukasz Dziedziak <lukidzi@gmail.com>