ci(gh): allow to run `build-test-distribute` on `workflow_dispatch` (#12149)
This is a part of the effort to run tests on CI for release branches on
schedule. As Github doesn't allow to run workflows on schedule from
branches other than default (master in our case), the best solution I
could find is to allow to run `build-test-distribute` workflow on
`workflow_dispatch` event and then in master on schedule send a REST
call to manually trigger these workflows on wanted release branches.

Signed-off-by: Bart Smykla <bartek@smykla.com>