ci(github): run scheduled CI for release branches using workflow_dispatch
We decided to run daily tests for release branches to monitor stability.
This change adds a new workflow in the master branch that triggers CI
on release branches using the `workflow_dispatch` event. The schedule
is:

- release-A: 11:00 AM PST / 20:00 CET / 03:00 CST
- release-B: 01:00 PM PST / 22:00 CET / 05:00 CST
- release-C: 03:00 PM PST / 00:00 CET / 07:00 CST
- release-D: 05:00 PM PST / 02:00 CET / 09:00 CST

Each branch's `build-test-distribute` workflow was already updated
earlier to support `workflow_dispatch`. The new master workflow uses the
GitHub API to trigger these workflows. This ensures we are able to check
branch stability with minimal changes and no interruptions to ongoing
work.

Signed-off-by: Bart Smykla <bartek@smykla.com>