feat(oapi): add mesh and meshgateway to oapi spec (#12272)
## Motivation

We want to be able to mange these resources using Terraform and need
OpenAPI spec to be able to generate that. Read the MADR if you're
interested in more details:
https://docs.google.com/document/d/1Ewe0yRx-vvAuQn5CWRFhJbBPLzdbBQ2wsQosZXk4Go8/
.

<!-- Why are we doing this change -->

## Implementation information

All described in the madr -
https://docs.google.com/document/d/1Ewe0yRx-vvAuQn5CWRFhJbBPLzdbBQ2wsQosZXk4Go8/edit?tab=t.0#heading=h.9m1k72xxi21x
.

<!-- Explain how this was done and potentially alternatives considered
and discarded -->

## Supporting documentation

<!-- Is there a MADR? An Issue? A related PR? -->

Closes https://github.com/kumahq/kuma/issues/11507

<!--
> Changelog: skip
-->
<!--
Uncomment the above section to explicitly set a [`> Changelog:` entry
here](https://github.com/kumahq/kuma/blob/master/CONTRIBUTING.md#submitting-a-patch)?
-->

---------

Signed-off-by: Charly Molter <charly.molter@konghq.com>
Signed-off-by: slonka <slonka@users.noreply.github.com>
Co-authored-by: Charly Molter <charly.molter@konghq.com>