perf(xds): `add x-kuma-tags` conditionally (#11076)
We use `x-kuma-tags` internally. We add this to outbound HTTP listeners
and we remove them on the inbound listeners. Currently, we use this
header for 3 things
1. Match sources in `FaultInjection` policy
2. Match from in `MeshFaultInjection` policy
3. Match sources in `RateLimit` policy

I'd argue that in 99% cases you don't need it then.
Just a simple app on k8s looks like this:
```json
"request_headers_to_add": [
  {
    "header": {
      "key": "x-kuma-tags",
      "value": "&app=redis&&k8s.kuma.io/namespace=kuma-demo&&k8s.kuma.io/service-name=redis&&k8s.kuma.io/service-port=6379&&kubernetes.io/hostname=k3d-kuma-server-0&&kuma.io/protocol=tcp&&kuma.io/service=redis_kuma-demo_svc_6379&&kuma.io/zone=default&&pod-template-hash=686bbbb5c8&"
    }
  }
]
```
This is quite a lot of data to put on every single request between
Envoys.
Additionally, that's quite a lot of data to put on every single outbound
in XDS config. Removing this slims down the XDS config.

Long term this header should be deleted and we should extract this from
the certificate on the inbound side.

Signed-off-by: Jakub Dyszkiewicz <jakub.dyszkiewicz@gmail.com>