feat(inspect): api for retrieving service hostnames (#11865)
## Motivation

We want to see hostnames of Mesh*Service in the Global CP.

## Implementation information

While we could add new KDSService to retrieve stats of a resource in
specific zones and then aggregate this an easier (less expensive to
execute, faster, don't depend on the zone) option is to calculate the
hostname. We have all the resources in global CP already (because
HostnameGenerator from zone is synced to global for visibility). It's
just a matter of relabeling properly.

## Supporting documentation

Fix https://github.com/kumahq/kuma/issues/10618

---------

Signed-off-by: Jakub Dyszkiewicz <jakub.dyszkiewicz@gmail.com>