fix(xds): ignore watchdog error on context cancelled (#12664)
## Motivation

We are producing logs that we should ignore

## Implementation information

I made a change to ignore OnError callback in case of ctx is cancelled.
Other option would be to change log level from ERROR to INFO dynamically
in OnError callback implementation, but we would need to repeat this
across all watchdog implems. I'm not sure we care about this error. The
intention of already existing `!errors.Is(err, context.Canceled)` seems
to be to ignore such errors.

While this is fix, I don't think it's worth to backport it. It's not a
stability issue, it's just a slight noise in logs.

## Supporting documentation

Fix #12663 

---------

Signed-off-by: Jakub Dyszkiewicz <jakub.dyszkiewicz@gmail.com>