test(meshtimeout): e2e test that checks golden envoy configs (#12650)
## Motivation

When we test `BuildRules` we pass MeshTimeout policies as input and
check `outbound.ResourceRules` Go struct.

When we test MeshTimeout plugin we pass `outbound.ResourceRules` Go
struct as input and check envoy configs.

We need an easy way to pass MeshTimeout policies and check envoy
configs. It's sort of an integration test that involves 2 packages. The
easiest way to do that is to add another universal e2e test with golden
files. This way we have running Kuma CP and we basically test it as a
black box.

I managed to find 2 bugs using these tests:
* https://github.com/kumahq/kuma/issues/12648
* https://github.com/kumahq/kuma/issues/12649

## Implementation information

We check DPP `config` (marshalled Snapshot in Kuma CP) instead of the
actual Envoy config dump for a few reasons:
* no need to add logic that waits while config is propagated on DPP
* less setup-specific values we have to reduct to have stable golden
file (only IP addresses)
* we can use `kuma.io/effect: shadow` and quickly review the `diff`
policies introduces

---------

Signed-off-by: Ilya Lobkov <ilya.lobkov@konghq.com>