feat(oapi): generate openapi schema for dynamically typed backend confs (#12680)
## Motivation

To quote @mheap:

```
Dynamic schemas are painful in Terraform.
We do all kinds of things in the Gateway spec to avoid them as much as we can. 
Without a schema, Terraform does not know which fields are required, what the defaults are etc.
It the server adds any defaults and your manifest does not contain those values, you will get a diff output on terraform plan every time it's run.
With schemas, Terraform can match response values against the schema and automatically remove them from the plan output.
```

## Implementation information

I put the types and mapping manually in the generator. Some parts are
made public to be overwritten in
https://github.com/Kong/kong-mesh/pull/7363

## Supporting documentation

xrel https://github.com/Kong/kong-mesh/issues/7201

---------

Signed-off-by: slonka <slonka@users.noreply.github.com>